requirements.txt
ariadne
uvicorn==0.10.8             #default from tiangolo/uvicorn-gunicorn:python3.7 is 0.9.0
gunicorn==20.0.4            #default from tiangolo/uvicorn-gunicorn:python3.7 is 19.9.0
asgi-lifespan==0.6.0        #latest is 1.0.0 and is not backward compatible
python-dotenv==0.10.3
requests
graphqlclient==0.2.4
